Use the Libraries module for including jStorage
Instead of using a path, I modified the code to use the Libraries
module to include and require the jStorage library. A warning will
show up in Reports->Status and an error will show up on the page
if the module cannot find the error. Looks for the library in
"libraries/jstorage" in the sites folder (such as sites/all).