Separated the admin functions into files and added AJAX functionality
The splashify.admin.inc file was getting pretty large, since it
contained all of the forms in the admin area. I was working on
adding AJAX functionality to the "How" tab when it occurred to me
that it would be much simpler if things were in separate files.

I also added AJAX functionality to the "How" tab that shows/hide
the window size box only if that value should be entered.