Minimize display mode features for mobile devices.
Under How: Mobile: Splash Display Mode...there was the same options as there
is for the desktop settings. After thinking about it, it doesn't seem practical
to use a lightbox or open a popup window for the mobile splash. So I took the
other options out and disabled the selection (so it always has Redirect
selected)...for now. Before options are added here, I would like to see
how people use this part of the module for mobile devices.