Fixed lightbox bug and cleaned up code
Discovered a bug that was happening when using the lightbox mode with
the sequentional path option set. It was skipping a page because the JS
code in the Drupal.Behavior call was being hit multiple times. Wrapping
the code in $(window).local() fixed the problem.

Updated the CHANGELOG and applied a few other small tweaks to the
hook_ini() function.