Issue #[1438874]: Validate path values in the admin.
On the "What" tab, we check to make sure each path is either an
alias or the source url for both the desktop paths and mobile
paths (when sequence or random is selected for the mode). A
similar thing is done on the "Where" tab, except we don't accept
alias paths. This makes it easier to work with, since we can
directly compare that value against $_GET['q'], without having to
figure out if it is an alias or not.