Fixed random splash page with Page Caching On and more Testing
I discovered that the new JS was incomplete because it did not
handle the option for displaying a random splash page (like it
was doing with the sequential option). I added this new JS code
and also moved the JS around a little bit. I also cleaned up
some of the PHP code.

Fixed a bug in the JS code with the sequence option in IE7 and
IE8. indexOf() doesn't exist in these browsers, so replaced
with the jQuery equivalent.

I also did more testing in IE7, IE8 and IE9 to make sure the
new JS code worked in all cases.