Took the JS out of Drupal Behaviors
The way the JS was written before was making the system slow. I
reworked some of the JS so that it is faster and fixed a few
conflicts with Colorbox. I discovered that having the ifrome
mode off caused the colorbox call to be hit multiple times. With
the ifrome option on, the browser referrer is set to the site and
is not hit again. I'm going to remove this option from the admin.