Fixed Redirect Flicker
When using the redirect option, you would briefly see the page
they were loading before the user was redirected to the splash
page. To work around this, when the DOM finishes loading,
we hide the whole html page (only for the redirect option) so
that we can redirect the user to the splash page.

There is somewhat of a blank page flash before the splash is
displayed, but it is a lot less obvious than what it was doing
before and should suffice.

It does all of this while still being search engin friendly. If
a search engine bot goes to the page, they will not hit the
splash page (like before).

For none redirect mode, this update should effect how the splash
page shows up.