Tweaked the JS Code for Detecting When on Splash Page
Instead of sending a query parementer to the splash page url,
we now detect if we are on the splash page by looking at a JS
array. This prevents the need to have to read query
parameters in the JS.

I also tried testing the code to now use the iframe option for
the colorbox. But the problem is that colorbox loads this
data through AJAX and we don't have any way of knowing it is
coming from this type of request. This results into an
infinite loop.