Issue #1854832 by paul.michalet,chrisroane: Add referrer check config and fix IE bug
Added an option under the When tab in the admin that disables the
referrer check. This will allow a site to display a splash page
on specific pages, even with visitors that are already on the site.

While testing this in IE7 + IE8, I encountered a bug in the JS
that was using indexOf (which won't work in IE8 and below).